Dynamic: Avoid copying of constant strings without backslash escapes
It appears OK for dynamic_Demangle() to return a pointer to its input
string, rather than to a copy, because the only uses are:

1. In dynamic_parser.c, on pointers to config file lines, and we pre-read
the entire config file (and includes) into a linked list of separate line
structs, and we don't de-allocate those.
2. In dynamic_compiler.c: #define APP_CFUNC(N), we only pass the returned
pointer to dyna_helper_appendn(), which does not store the pointer.

Mostly fixes #5465 (memory leak)